conda-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_py-xgboost-mutex=2.0=cpu_0
absl-py=0.9.0=pypi_0
astor=0.8.1=pypi_0
attrs=19.3.0=pypi_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.0=pypi_0
bzip2=1.0.8=h516909a_2
ca-certificates=2020.4.5.1=hecc5488_0
cachetools=4.0.0=pypi_0
cairo=1.16.0=h18b612c_1001
catboost=0.20.2=py37_0
certifi=2020.4.5.1=py37hc8dfbb8_0
chardet=3.0.4=pypi_0
click=7.1.1=py_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py_2
cython=0.29.15=pypi_0
dbus=1.13.6=he372182_0
decorator=4.4.1=py_0
defusedxml=0.6.0=pypi_0
entrypoints=0.3=pypi_0
expat=2.2.9=he1b5a44_2
ffmpeg=4.1.3=h167e202_0
flake8=3.7.9=pypi_0
flask=1.1.1=py_1
fontconfig=2.13.1=he4413a7_1000
freetype=2.9.1=h8a8886c_1
gast=0.2.2=pypi_0
gettext=0.19.8.1=hc5be6a0_1002
giflib=5.2.1=h516909a_2
glib=2.58.3=py37h6f030ca_1002
gmp=6.2.0=he1b5a44_2
gnutls=3.6.5=hd3a4fd2_1002
google-auth=1.11.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.1.8=pypi_0
graphite2=1.3.13=hf484d3e_1000
grpcio=1.26.0=pypi_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
h5py=2.10.0=nompi_py37h513d04c_102
harfbuzz=2.4.0=h37c48d4_1
hdf5=1.10.5=nompi_h3c11f04_1104
icu=58.2=hf484d3e_1000
idna=2.8=pypi_0
importlib-metadata=1.5.0=pypi_0
intel-openmp=2019.4=243
ipykernel=5.1.3=py37h39e3cac_1
ipython=7.11.1=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=pypi_0
itsdangerous=1.1.0=py37_0
jasper=1.900.1=h07fcdf6_1006
jedi=0.15.2=py37_0
jinja2=2.11.1=py_0
joblib=0.14.1=py_0
jpeg=9c=h14c3975_1001
jsonschema=3.2.0=pypi_0
jupyter_client=5.3.4=py37_0
jupyter_core=4.6.1=py37_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
kiwisolver=1.1.0=py37hc9558a2_0
lame=3.100=h14c3975_1001
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1005
liblapack=3.8.0=14_mkl
liblapacke=3.8.0=14_mkl
libpng=1.6.37=hbc83047_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=2.32.1=h14c3975_1000
libwebp=1.0.2=h56121f0_5
libxcb=1.13=h14c3975_1002
libxgboost=0.81=hf484d3e_1000
libxml2=2.9.9=hea5a465_1
markdown=3.1.1=pypi_0
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
mccabe=0.6.1=pypi_0
mistune=0.8.4=pypi_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
nbconvert=5.6.1=pypi_0
nbformat=5.0.4=pypi_0
ncurses=6.1=he6710b0_1
nettle=3.4.1=h1bed415_1002
networkx=1.11=py37_1
ninja=1.9.0=py37hfd86e86_0
nltk=3.4.4=py_0
notebook=6.0.3=pypi_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
opencv=4.1.1=py37ha799480_1
opencv-python=4.2.0.32=pypi_0
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1g=h516909a_0
opt-einsum=3.1.0=pypi_0
pandas=0.25.3=py37he6710b0_0
pandocfilters=1.4.2=pypi_0
parso=0.5.2=py_0
pcre=8.44=he1b5a44_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_0
pillow=7.0.0=py37hb39fc2d_0
pip=20.0.2=pypi_0
pixman=0.38.0=h516909a_1003
plotly=4.6.0=pyh9f0ad1d_0
prometheus-client=0.7.1=pypi_0
prompt_toolkit=3.0.2=py_0
protobuf=3.11.2=pypi_0
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=py37_0
py-xgboost=0.81=py37hf484d3e_1000
pyarrow=0.16.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.0=py37h516909a_1001
pycodestyle=2.5.0=pypi_0
pyflakes=2.1.1=pypi_0
pygments=2.5.2=py_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37hcca6a23_4
pyrsistent=0.15.7=pypi_0
python=3.7.6=h0371630_2
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.4.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pytz=2019.3=py_0
pyzmq=18.1.0=py37he6710b0_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
requests=2.22.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
retrying=1.3.3=py_2
rsa=4.0=pypi_0
scikit-learn=0.22.1=py37hd81dba3_0
scipy=1.4.1=pypi_0
send2trash=1.5.0=pypi_0
setuptools=44.0.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.13.0=py37_0
sqlite=3.30.1=h7b6447c_0
tensorboard=2.1.0=pypi_0
tensorflow=2.1.0=pypi_0
tensorflow-estimator=2.1.0=pypi_0
tensorflow-gpu=2.1.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
tk=8.6.8=hbc83047_0
torchvision=0.5.0=py37_cu101
tornado=6.0.3=py37h7b6447c_0
tqdm=4.45.0=pyh9f0ad1d_0
traitlets=4.3.3=py37_0
urllib3=1.25.8=pypi_0
wcwidth=0.1.7=py37_0
webencodings=0.5.1=pypi_0
werkzeug=0.16.1=pypi_0
wheel=0.33.6=py37_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.11.2=pypi_0
x264=1!152.20180806=h14c3975_0
xgboost=0.81=py37hf484d3e_1000
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
zeromq=4.3.1=he6710b0_3
zipp=2.1.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
